+++ packages/config/src/events.js
@@ -1,14 +1,9 @@
-const CHANGE = "change";
+export const CHANGE = "change";
 
-function addEventListener(listener, eventName, eventEmitter) {
+export function addEventListener(listener, eventName, eventEmitter) {
   const removeCallback = () => {
     eventEmitter.removeListener(eventName, listener);
   };
   eventEmitter.on(eventName, listener);
   return removeCallback;
 }
-
-module.exports = {
-  addEventListener,
-  CHANGE,
-};