+++ packages/config/test/.eslintrc.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  parser: "@babel/eslint-parser",
+  parserOptions: {
+    sourceType: "module",
+    allowImportExportEverywhere: true,
+    requireConfigFile: false,
+  },
+};