+++ packages/config/test/specs/index.spec.js
@@ -1,0 +1,8 @@
+import Config from "../../src/Config";
+import index from "../../src/index";
+
+describe("index", () => {
+  it("should export Config", () => {
+    expect(index).toBe(Config);
+  });
+});