+++ packages/config/CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
+## [Unreleased]
+
+### Changed
+- refactor: In CJS, .default has to be required
+- refactor: Migrate to TypeScript
+
 ## [1.4.0] - 2022-09-01
 
 ### Added